Use double for blending mode simulations to have consistent behaviour on JS

+++ korim/src/commonMain/kotlin/com/soywiz/korim/internal/MathExt.kt
@@ -11,8 +11,8 @@ internal fun packIntUnchecked(r: Int, g: Int, b: Int, a: Int): Int {
 internal fun packIntClamped(r: Int, g: Int, b: Int, a: Int): Int =
     packIntUnchecked(r.clamp0_255(), g.clamp0_255(), b.clamp0_255(), a.clamp0_255())
 
-internal fun d2i(v: Double): Int = ((v.toFloat()).clamp01() * 255).toInt()
-internal fun f2i(v: Float): Int = ((v).clamp01() * 255).toInt()
+internal fun d2i(v: Double): Int = (v.clamp01() * 255).toInt()
+internal fun f2i(v: Float): Int = (v.clamp01() * 255).toInt()
 
 internal fun floorCeil(v: Double): Double = if (v < 0.0) ceil(v) else floor(v)
